Done homework "06_recyclerview"  task "RecyclerView" version 2 move PostAdapter to MainActivity
