+++ app/src/main/java/ru/netology/nmedia/PostsAdapter.kt
@@ -1,4 +1,4 @@
-package ru.netology.nmedia.data.impl
+package ru.netology.nmedia
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
